+++ srcs/custom1.c
@@ -6,7 +6,7 @@
 /*   By: dchirol <dchirol@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/06/05 20:05:34 by dchirol           #+#    #+#             */
-/*   Updated: 2017/06/05 20:37:54 by dchirol          ###   ########.fr       */
+/*   Updated: 2017/06/05 20:55:29 by dchirol          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -66,14 +66,14 @@ void			custom1_calc(t_env *e, t_pnt *min)
 		e->i++;
 	}
 	if (e->i != ITER_MAX)
-		mlx_pixel_put(e->image, e->win, e->x, e->y, ft_getcolor((float)e->i
+		mlx_pixel_put(e->image, e->win, e->x, e->y, ft_getcolor((double)e->i
 		+ 1 - log(log(4) / log(z.re * z.re + z.im * z.im)) / log(2)));
 }
 
 void			ft_custom1(t_env *e, t_pnt min, t_pnt max)
 {
-	float	step;
-	float	tmp;
+	double	step;
+	double	tmp;
 
 	mlx_clear_window(e->mlx, e->win);
 	step = (max.x - min.x) / WINX;
